Task 10
I made use of the “push down” function to move the “available” variable
and its functions to the sub classes. This caused an issue, so I made
use of the “pull up” refactor button and chose every cell subclass, so
that it would correctly pull up the “available” variable and supporting
functions to the superclass.