Task 13
Worked on refactoring with the “extract a variable” command. It allowed
me to take code that was repeated often and assign it to a variable
once. Then, just reference that one variable instead of calling a
method over and over.  Specifically, “cell.getColorGroup()” was
assigned to the variable “colorGroup”. The variable “colorGroup” was
used for references to values from there.