Task 15
I used the bad smell add-on to find pieces of code that could make use
of some refactoring.  Once it showed me pieces of code that need it, I
refactored my code.  I created a setupHelp method in GainMoneyCardTest.
Also, i created a method called setPlayerName, which focuses on only
naming a play. Then, it allows the code in the original method to flow
better.